feat: implements docker compose to manage application and database
application and database (which is now postgrees) are being managed by docker compose and the tests are now using a postgres database running in a test container

+++ tests/test_security.py
@@ -1,50 +1,43 @@
 from http import HTTPStatus
 
+import pytest
 from jwt import decode
 
 from ..fast_zero.security import create_access_token, settings
 
 
-# @pytest.mark.skip(reason='Already tested')
+@pytest.mark.skip(reason='Already tested')
 def test_create_access_token():
     data = {'secret': 'Meu segredo'}
     token = create_access_token(data)
 
-    decoded_data = decode(
-        token, settings.SECRET_KEY, algorithms=[settings.ALGORITHM]
-    )
+    decoded_data = decode(token, settings.SECRET_KEY, algorithms=[settings.ALGORITHM])
 
     assert data['secret'] == decoded_data['secret']
     assert 'exp' in decoded_data
 
 
-# @pytest.mark.skip(reason='Already tested')
+@pytest.mark.skip(reason='Already tested')
 def test_jwt_invalid_token(client):
-    response = client.delete(
-        '/users/1', headers={'Authorization': 'Bearer invalid_token'}
-    )
+    response = client.delete('/users/1', headers={'Authorization': 'Bearer invalid_token'})
 
     assert response.status_code == HTTPStatus.UNAUTHORIZED
     assert response.json() == {'detail': 'Could not validate credentials'}
 
 
-# @pytest.mark.skip(reason='Already tested')
+@pytest.mark.skip(reason='Already tested')
 def test_get_current_user_user_not_found(client):
     token = create_access_token({'sub': 'not_found'})
-    response = client.delete(
-        '/users/1', headers={'Authorization': f'Bearer {token}'}
-    )
+    response = client.delete('/users/1', headers={'Authorization': f'Bearer {token}'})
 
     assert response.status_code == HTTPStatus.UNAUTHORIZED
     assert response.json() == {'detail': 'Could not validate credentials'}
 
 
-# @pytest.mark.skip(reason='Already tested')
+@pytest.mark.skip(reason='Already tested')
 def test_get_current_user_user_found(client):
     token = create_access_token({})
-    response = client.delete(
-        '/users/1', headers={'Authorization': f'Bearer {token}'}
-    )
+    response = client.delete('/users/1', headers={'Authorization': f'Bearer {token}'})
 
     assert response.status_code == HTTPStatus.UNAUTHORIZED
     assert response.json() == {'detail': 'Could not validate credentials'}
