+++ tests/test_users.py
@@ -1,9 +1,11 @@
 from http import HTTPStatus
 
+import pytest
+
 from ..fast_zero.schemas import UserPublic
 
 
-# @pytest.mark.skip(reason='Already tested')
+@pytest.mark.skip(reason='Already tested')
 def test_create_user(client):
     response = client.post(
         '/users',
@@ -22,7 +24,7 @@ def test_create_user(client):
     }
 
 
-# @pytest.mark.skip(reason='Already tested')
+@pytest.mark.skip(reason='Already tested')
 def test_create_user_username_exists(client, user):
     response = client.post(
         '/users',
@@ -39,7 +41,7 @@ def test_create_user_username_exists(client, user):
     }
 
 
-# @pytest.mark.skip(reason='Already tested')
+@pytest.mark.skip(reason='Already tested')
 def test_create_user_email_exists(client, user):
     response = client.post(
         '/users',
@@ -56,14 +58,14 @@ def test_create_user_email_exists(client, user):
     }
 
 
-# @pytest.mark.skip(reason='Already tested')
+@pytest.mark.skip(reason='Already tested')
 def test_read_users(client):
     response = client.get('/users')
     assert response.status_code == HTTPStatus.OK
     assert response.json() == {'users': []}
 
 
-# @pytest.mark.skip(reason='Already tested')
+@pytest.mark.skip(reason='Already tested')
 def test_read_users_with_user(client, user):
     user_schema = UserPublic.model_validate(user).model_dump()
     response = client.get('/users')
@@ -71,7 +73,7 @@ def test_read_users_with_user(client, user):
     assert response.json() == {'users': [user_schema]}
 
 
-# @pytest.mark.skip(reason='Already tested')
+@pytest.mark.skip(reason='Already tested')
 def test_read_user(client, user):
     user_schema = UserPublic.model_validate(user).model_dump()
     response = client.get('/users/1')
@@ -79,7 +81,7 @@ def test_read_user(client, user):
     assert response.json() == user_schema
 
 
-# @pytest.mark.skip(reason='Already tested')
+@pytest.mark.skip(reason='Already tested')
 def test_read_user_not_found(client, user):
     response = client.get('/users/2')
     assert response.status_code == HTTPStatus.NOT_FOUND
@@ -88,7 +90,7 @@ def test_read_user_not_found(client, user):
     }
 
 
-# @pytest.mark.skip(reason='Already tested')
+@pytest.mark.skip(reason='Already tested')
 def test_update_user(client, user, token):
     response = client.put(
         f'/users/{user.id}',
@@ -108,7 +110,7 @@ def test_update_user(client, user, token):
     }
 
 
-# @pytest.mark.skip(reason='Already tested')
+@pytest.mark.skip(reason='Already tested')
 def test_update_user_not_current_user(client, user, other_user, token):
     response = client.put(
         f'/users/{other_user.id}',
@@ -125,7 +127,7 @@ def test_update_user_not_current_user(client, user, other_user, token):
     }
 
 
-# @pytest.mark.skip(reason='Already tested')
+@pytest.mark.skip(reason='Already tested')
 def test_update_user_integrity_error(client, user, other_user, token):
     response = client.put(
         f'/users/{user.id}',
@@ -142,22 +144,18 @@ def test_update_user_integrity_error(client, user, other_user, token):
     }
 
 
-# @pytest.mark.skip(reason='Already tested')
+@pytest.mark.skip(reason='Already tested')
 def test_delete_user(client, user, token):
-    response = client.delete(
-        f'/users/{user.id}', headers={'Authorization': f'Bearer {token}'}
-    )
+    response = client.delete(f'/users/{user.id}', headers={'Authorization': f'Bearer {token}'})
     assert response.status_code == HTTPStatus.OK
     assert response.json() == {
         'message': 'User deleted successfully',
     }
 
 
-# @pytest.mark.skip(reason='Already tested')
+@pytest.mark.skip(reason='Already tested')
 def test_delete_user_not_found(client, other_user, token):
-    response = client.delete(
-        f'/users/{other_user.id}', headers={'Authorization': f'Bearer {token}'}
-    )
+    response = client.delete(f'/users/{other_user.id}', headers={'Authorization': f'Bearer {token}'})
     assert response.status_code == HTTPStatus.FORBIDDEN
     assert response.json() == {
         'detail': 'Not enough permissions',
